backend/requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database - Supabase
supabase==2.3.4
postgrest==0.16.1
realtime==1.0.2
storage3==0.7.0
gotrue==2.4.0

# Google Cloud Services
google-cloud-speech==2.24.1
google-cloud-translate==3.15.0
google-generativeai==0.4.0

# WebSocket & Real-time
websockets==12.0
python-socketio==5.11.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.6.0
pydantic-settings==2.1.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.3

# Audio Processing
pydub==0.25.1
soundfile==0.12.1
numpy==1.26.3

# Utilities
python-dotenv==1.0.1
redis==5.0.1
celery==5.3.6
structlog==24.1.0

# Video Conference SDKs
zoomus==1.1.6

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
httpx==0.26.0

